[#37] Move auditor logic to new module
