+++ src/auditor/mod.rs
@@ -1,0 +1,7 @@
+//! `auditor` is a module containing implementations of the auditors.
+
+pub mod common;
+pub mod metric;
+pub mod opcode;
+pub mod retransmission;
+pub mod util;
